fix pytest memory usage
