+++ tests/python/ops/conftest.py
@@ -1,0 +1,52 @@
+# Copyright © Advanced Micro Devices, Inc. All rights reserved.
+#
+# MIT License
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+import os
+
+import pytest
+from tests.python.ops.dispatch_combine_test_utils import (
+    start_torch_dist_process_manager,
+)
+
+
+@pytest.fixture(scope="session")
+def torch_dist_process_manager():
+    """Single shared worker pool for all ops tests."""
+    manager = start_torch_dist_process_manager(world_size=8)
+    yield manager
+    manager.shutdown()
+
+
+@pytest.fixture(scope="session", autouse=True)
+def set_shmem_heap_size():
+    # Set shmem heap size for all tests in this directory.
+    # 32G is the minimum required by the large-token internode_v1 test
+    # (~17.7 GB per rank).  Using a session-scoped fixture (instead of
+    # pytest_configure) so the override is scoped to ops/ tests only and
+    # does not bleed into other test directories (e.g. shmem/) when the
+    # full suite is collected together.
+    prev = os.environ.get("MORI_SHMEM_HEAP_SIZE")
+    os.environ["MORI_SHMEM_HEAP_SIZE"] = "32G"
+    yield
+    if prev is None:
+        os.environ.pop("MORI_SHMEM_HEAP_SIZE", None)
+    else:
+        os.environ["MORI_SHMEM_HEAP_SIZE"] = prev
